requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0